requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aistudio-sdk==0.2.1
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
async-timeout==4.0.3
asynctest==0.13.0
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
bce-python-sdk==0.9.6
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.10
datasets==2.13.2
decorator==5.1.1
dill==0.3.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.103.2
filelock==3.12.2
fire==0.6.0
Flask==2.2.5
Flask-Babel==2.0.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==1.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.7
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lmdb==1.4.1
lxml==5.2.1
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.12.2
networkx==2.6.3
numpy==1.21.6
onnx==1.14.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddle-bfloat==0.1.7
paddle2onnx==1.1.0
paddlefsl==1.1.0
paddlenlp==2.5.2
paddleocr==2.6.1.2
paddlepaddle-gpu==2.4.2.post112
pandas==1.3.5
pdf2docx==0.5.8
Pillow==9.5.0
premailer==3.10.0
prettytable==3.7.0
protobuf==3.20.0
psutil==5.9.8
pyarrow==12.0.1
pybind11==2.12.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyMuPDF==1.20.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
rapidfuzz==3.4.0
rarfile==4.2
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.2.0
seqeval==1.2.2
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
starlette==0.27.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tool-helpers==0.1.1
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
visualdl==2.4.2
wcwidth==0.2.13
Werkzeug==2.2.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.15.0
